refactor(layer-separation): dto 파일을 도메인 패키지로 이동
