+++ src/main/java/roomescape/reservationtime/ReservationTimeController.java
@@ -8,8 +8,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.ReservationTimeRequestDTO;
-import roomescape.dto.ReservationTimeResponseDTO;
 
 @RestController
 @RequestMapping("/times")
